Refactored listing form to use custom components for form elements
